requirements.txt
streamlit>=1.45.1
PyMuPDF>=1.23.0
Pillow>=10.0.0
openai>=1.84.0 